+++ include/sel.h
@@ -18,12 +18,12 @@ void sel_hold(void);
 void sel_disable(void);
 void sel_print_index(void);
 uint8_t sel_move_check_dis(uint8_t caller, uint8_t arg);
-void sel_set_max_index(uint8_t num_tubes);
+void sel_set_max_index(uint16_t num_tubes);
 void sel_home_init(void);
-void sel_move_init(uint8_t destination_index);
+void sel_move_init(uint16_t destination_index);
 void sel_move_next_init(void);
 
-extern volatile uint8_t sel_index;
+extern volatile uint16_t sel_index;
 extern volatile uint8_t sel_move_after_dis_home[2];
 
 #endif
